Shell of login, company, homepage
Doesn't work right. Not sure how to get it going. Needs to open to login.html, and if they are a current user, login or if they need to create a company account, hit the "get started" button to move to company.html. Once they create the company or login they should move to homepage.html shell. I basically wrote a bunch of psudo code without the ability to test it.